backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

beanie==1.23.6
motor==3.3.2
pymongo==4.6.1

fastapi-users[beanie]==12.1.2
fastapi-users-db-beanie

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4


python-dotenv==1.0.1


textblob==0.17.1
pytrends==4.9.2
requests==2.31.0
praw==7.7.1
tweepy==4.14.0
bcrypt==4.0.1